requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.1.4
scikit-learn==1.4.0
joblib==1.3.2
sqlalchemy==2.0.25
python-dotenv==1.0.1
pydantic==2.6.3
numpy==1.26.4